Create CNN_LSTM_Autoencoder.py
Code for a CNN_LSTM Autoencoder. Currently consists of 4 Convolutional Encoder Blocks (32,64,128,256), then an LSTM Autoencoder block with a variable Hidden size param ( I tried using 16,32,64 and 128), and a Convolutional Decoder block with the same architecture as the encoder but in reverse. Included code to resume training from a saved model automatically, and log model architecture and metrics in Tensorboard.